+++ src/Printers.java
@@ -3,7 +3,7 @@ import java.util.HashSet;
 import java.util.List;
 
 public class Printers {
-    public void printMap(HashMap<Integer, List<Pair>> hashMap){
+    public void printMap(HashMap<Long, List<Pair>> hashMap){
         System.out.println(hashMap);
     }
